complated front projects
We did a project under the management of Professor Ali Pourmoman and Ehsan Peyman and our good team of 9 people
This is an almost big project. I am Rahela Shirazi. I worked in the front part of the project and my good teammate Tahira Safari and Mr. Morsad are working together.
We faced a lot of challenges, but we were able to solve it with the help of our project manager, and this is what happened.